+++ examples/pytorch/explicit_argument_type_annotation_mnist.py
@@ -5,7 +5,7 @@
 import torch
 import torch_xla.core.xla_model as xm
 import torch_xla.runtime as xr
-import tt_torch
+import tt_torch_backend as tt_torch
 
 from tests.torch.single_chip.models.mnist.cnn.dropout.model_implementation import (
     MNISTCNNDropoutModel,
