+++ examples/pytorch/serialization_example.py
@@ -10,7 +10,7 @@ import torch_xla.core.xla_model as xm
 import torch_xla.runtime as xr
 from torch_xla.experimental import plugins
 
-from tt_torch import parse_compiled_artifacts_from_cache_to_disk
+from tt_torch_backend import parse_compiled_artifacts_from_cache_to_disk
 
 xr.set_device_type("TT")
