+++ tests/torch/single_chip/ops/test_custom_ops.py
@@ -7,8 +7,8 @@ import torch_xla.core.xla_model as xm
 
 import pytest
 from infra.utilities.types import Framework
-import tt_torch.custom_ops
-import tt_torch
+import tt_torch_backend.custom_ops
+import tt_torch_backend
 from tests.infra.testers.single_chip.op.op_tester import OpTester, run_op_test
 
 # TODO: Record superset properties for these tests.
